+++ pages/_app.js
@@ -1,4 +1,4 @@
-import "../components/global.css";
+import "../neró/global.css";
 import { useEffect } from "react";
 import { useRouter } from "next/router";
 import * as Fathom from "fathom-client";
